更新

+++ xjhqreIOT-system/src/main/java/com/xjhqre/system/mapper/MenuMapper.java
@@ -64,5 +64,5 @@ public interface MenuMapper extends BaseMapper<Menu> {
      * @param parentId
      * @return
      */
-    Integer getMaxId(@Param("parentId") Long parentId);
+    Long getMaxId(@Param("parentId") Long parentId);
 }
